requirements.txt
discord.py==1.7.3
pandas==1.1.5
python-dotenv==0.19.0
networkx==2.5.1
decorator==4.4.2